scraper/requirements.txt
flask==3.1.3
flask-cors==6.0.2
yfinance==1.2.0
pandas==2.2.0
lxml==5.1.0
google-genai==0.3.0
python-dotenv==1.0.1
